+++ compiler/src/values/index.ts
@@ -3,7 +3,6 @@ export * from "./BaseValue";
 export * from "./LiteralValue";
 export * from "./StoreValue";
 export * from "./ObjectValue";
-export * from "./SenseableValue";
 export * from "./DestructuringValue";
 export * from "./LazyValue";
 export * from "./AssignmentValue";